final-project/requirements.txt
flask
mysql
mysql-connector
flask-mysqldb
